feat: an example on how to work with hash maps
Here actually I'm still starting this example. The main idea of this
example is run in a while true loop to CRUD countries and medals, to
have in the end a kind of Olympics medal standings. The idea is having
all this data stored in hash maps. Another thing that is being built is
a kind of paginator, to be able to print to the console more information
than fits to the screen and let the user read, without the need of
scrolling.